Submit Answer Functionality Added
The Submit Answer button now submits however many answers you have
create with the Create Answer UI as a user. It checks it against a key
in a text file. If you created too many or few solutions an alert
informs you that you have done just that. Otherwise, if you make one
solution that does not match the key an alert informs you to try again.
Otherwise, your answer is correct (since the input answer text boxes
already ensure nothing but numbers are entered) and an alert informs you
just of this. The Submit Answer button works as an AJAX post from JQuery
to the ASP code behind file, which has two WebMethods, and the button
accesses the Grading method with the answers passed from the user's
answer boxes. There is also an optional function and method in both the
aspx and the code behind file to change the solution key in the text
file. I would best advise you use it inside the first parts of the code
behind file where you generate your random matrix.

You guys did fantastic work while I was playing in the Civic Concert
last Friday. I thought it was only fair that I added substantial work of
my own. As for the text file with the solutions, the file path will need
to be changed on another computer, but this web site or app would
ideally be deployed from a server anyway. Right?

Feel free to text or email me if you have questions. Thanks for putting
up with me to this point.