Added free variables window
We now have the ability to have the user turn solutions into free
variables without crashing the server. I also added a string list and
string array to the grading methods and to the page load of the code
behind page. Someone can edit the matrix generation code to allow the
new string list to be populated with an "f" if a free solution is
generated from the random matrix. That's the next step...