Split user documents and added javascript check
I split the user documents into two documents, one for students and one
for instructors. In addition, I added the javascript check in input
boxes for minus signs and fraction bars because I felt like that was
going to cause problems. I also commented out the html that displayed
the dynamic text on the default page.